chore: Code Refactor
