+++ src/types/index.ts
@@ -4,3 +4,4 @@ export * from './LintConfig'
 export * from './LintRule'
 export * from './LintRuleType'
 export * from './Severity'
+export * from './Macro'
